functions/arcgis/requirements.txt
requests==2.24.0
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-firestore==1.6.0
googleapis-common-protos==1.51.0
zulu==1.2.0